+++ Sources/Sentry/include/SentryUIApplication.h
@@ -1,9 +1,8 @@
-#import <Foundation/Foundation.h>
 #import "SentryDefines.h"
-
+#import <Foundation/Foundation.h>
 
 #if SENTRY_HAS_UIKIT
-#import <UIKit/UIKit.h>
+#    import <UIKit/UIKit.h>
 
 NS_ASSUME_NONNULL_BEGIN